Move dispatcher tests out of func-test
