+++ haskell-ide-engine.cabal
@@ -176,8 +176,11 @@ test-suite dispatcher-test
   hs-source-dirs:      test/dispatcher
   main-is:             Main.hs
   build-depends:       base
+                     , aeson
                      , containers
                      , data-default
+                     , directory
+                     , filepath
                      , haskell-lsp
                      , haskell-ide-engine
                      , hie-test-utils
@@ -185,6 +188,7 @@ test-suite dispatcher-test
                      , hspec
                      , stm
                      , text
+                     , unordered-containers
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N -Wall -Wredundant-constraints
   if flag(pedantic)
      ghc-options:      -Werror
@@ -220,7 +224,6 @@ test-suite func-test
   main-is:             Main.hs
   other-modules:       CompletionSpec
                      , CodeActionsSpec
-                     , DispatchSpec
                      , DeferredSpec
                      , ReferencesSpec
                      , RenameSpec
